Placing the logic of the list in a separate file
